Move l1a data validation helper funtions to a separate helpers directory
